[DATASET][ADD] Add load_dataset function to load all csv files in one dataframe
