+++ src/main.py
@@ -6,6 +6,7 @@ from config import load_config
 from data.dataset import (
     build_dataset,
     create_pytorch_dataset,
+    load_dataset,
     plot_sym_train_test,
 )
 
@@ -18,7 +19,10 @@ if __name__ == "__main__":
     if args.build_dataset:
         build_dataset()
 
-    (train, test) = create_pytorch_dataset(
-        "ADAUSDT", client.CLIENT.KLINE_INTERVAL_1MINUTE
-    )
-    plot_sym_train_test(train=train, test=test)
+    df = load_dataset()
+    print(df.memory_usage())
+
+    # (train, test) = create_pytorch_dataset(
+    #     "ADAUSDT", client.CLIENT.KLINE_INTERVAL_1MINUTE
+    # )
+    # plot_sym_train_test(train=train, test=test)
